+++ CHANGELOG.rst
@@ -6,6 +6,8 @@
   Update Swagger UI from 2.0.17 to 2.0.18
 * `#433 <https://github.com/xitrum-framework/xitrum/issues/433>`_
   Update D3.js from 3.4.8 to 3.4.11
+* `#435 <https://github.com/xitrum-framework/xitrum/issues/435>`_
+  Update jQuery Validation from 0.12.0 to 0.13.0
 
 3.16: